+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Unreleased
 
+  * Add `SdpApi::stop_media()` to stop an existing m-line #941
   * Always use ActPass when sending an SDP offer (RFC 8842 Section 5.2/5.5) #893
   * Fix frame drops in Firefox due missing RTX SSRC #940
   * Drop stale depacketizer state on stream pause and restore paused timestamp repair #929